requirements.txt
streamlit==1.28.0
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.1
pillow==10.0.0
